+++ src/lib/graph/sz-graph.component.ts
@@ -184,7 +184,9 @@ export class SzGraphComponent implements OnInit, OnDestroy {
   public filterShowMatchKeys: string[];
   public filterShowMatchKeyTokens: Array<SzMatchKeyTokenComposite>;
   private _showMatchKeysFilters: boolean = true;
-  private _showMatchKeyTokenFilters: boolean = true;
+  // temporarily switched to false since somethings
+  // broken (see https://github.com/Senzing/sdk-graph-components/issues/154)
+  private _showMatchKeyTokenFilters: boolean = false;
   private _showMatchKeyControl: boolean = true;
   @Input() public showCoreMatchKeyTokenChips: boolean       = false;
   @Input() public showExtraneousMatchKeyTokenChips: boolean = true;